Finished commenting the .cpp files.

+++ template.cpp
@@ -388,7 +388,10 @@ int main(int argc, char** argv)
 				game->KeyUp(event.key.keysym.scancode);
 				break;
 			case SDL_MOUSEMOTION:
-				game->MouseMove(vec2(event.motion.x, event.motion.y));
+				game->MouseMove(vec2(
+					static_cast<float>(event.motion.x), 
+					static_cast<float>(event.motion.y)
+				));
 				break;
 			case SDL_MOUSEBUTTONUP:
 				game->MouseUp(event.button.button);
